Check scriptcs_packages.config exists before deleting

+++ src/ScriptCs.Octokit.Sample/RunIntegrationTest.ps1
@@ -1,6 +1,8 @@
 ﻿$ErrorActionPreference = "Stop";
 scriptcs -cl
-rm scriptcs_packages.config
+if (Test-Path scriptcs_packages.config) {
+    rm scriptcs_packages.config
+}
 scriptcs -i ScriptCs.Octokit -P $args[0] -pre
 scriptcs sample.csx
 exit $LASTEXITCODE
